+++ unreal/AegisWar/Source/AegisWar/Public/WarWorldEditWidget.h
@@ -21,6 +21,7 @@ private:
     void SelectNearest();
     void PlaceTemplate(FName TemplateId);
     void SnapSelected();
+    void SetSelectedComponent(int32 Field, double Value);
     void EditSelected(FVector Offset, double Yaw = 0, double Scale = 1, bool bToggleHidden = false);
     FName Selected;
     FString Search;